IOS-00024 Read WorkoutType Data IOS-00024C Strength Scores
