+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scroe/StrengthScore/Feature/StrengthScoreFeature+Action.swift
@@ -13,23 +13,26 @@ extension StrengthScoreFeature {
     
     @CasePathable
     enum Action: ViewAction {
+        
         // MARK: - Actions
         
+        /// Represents an action that updates the grouped workouts data.
         case groupedWorkouts
-        case findBestWorkout
         
         // MARK: - View Actions
         
+        /// Represents an action triggered by a view event.
         case view(View)
         
-        ///
         enum View {
-            
+            /// Indicates that the view has appeared.
             case viewDidAppear
+            
         }
         
         // MARK: - Destination
         
+        /// Represents navigation-related actions within the feature.
         case destination(PresentationAction<Destination.Action>)
     }
