+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scroe/StrengthScore/Feature/StrengthScoreFeature+Destination.swift
@@ -13,7 +13,7 @@ extension StrengthScoreFeature {
     
     @Reducer
     enum Destination {
-        
+      
     }
     
 }
