+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scroe/StrengthScore/Feature/StrengthScoreFeature+State.swift
@@ -23,9 +23,6 @@ extension StrengthScoreFeature {
         ///
         var groupedWorkoutsData: [GroupedWorkouts]?
         
-        ///
-        var bestWorkout: WorkoutStrength?
-        
         // MARK: - Destination
         
         /// destination from `StrengthScoreFeature`
