+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scroe/StrengthScore/Utilities/StrengthScoreService.swift
@@ -1,39 +1,24 @@
 //
-//  DefaultStrengthScoreService.swift
+//  StrengthScoreService.swift
 //  StepTrackerTCA
 //
-//  Created by Sebastian Sciuba on 10/03/2025.
+//  Created by Sebastian Sciuba on 11/03/2025.
 //
 
 import Foundation
 
+/// Protocol defining services related to strength assessment based on workouts.
 protocol StrengthScoreService {
     
-    func findBestWorkout(from workouts: [WorkoutStrength]) -> WorkoutStrength?
+    /// Finds the best workout based on the highest value.
+    ///
+    /// - Parameter workouts: A list of strength workouts.
+    /// - Returns: The best workout based on the highest value.
+    func findBestWorkout(from workouts: [WorkoutStrength]) -> WorkoutStrength
+
+    /// Groups workouts by movement type.
+    ///
+    /// - Parameter data: A list of strength workouts to be grouped.
+    /// - Returns: A list of grouped workouts, each containing the best workout for that group.
     func groupWorkoutsByMovement(_ data: [WorkoutStrength]) -> [GroupedWorkouts]
 }
-
-final class DefaultStrengthScoreService: StrengthScoreService {
-    
-    func findBestWorkout(from workouts: [WorkoutStrength]) -> WorkoutStrength? {
-        return workouts.max(by: { Double($0.value) ?? 0.0 < Double($1.value) ?? 0.0 })
-    }
-    
-    func groupWorkoutsByMovement(_ data: [WorkoutStrength]) -> [GroupedWorkouts] {
-        var groupedWorkouts: [GroupedWorkouts] = []
-        
-        for movement in StrengthMovement.allCases {
-            let workoutsForMovement = data.filter { $0.movement == movement }
-            if !workoutsForMovement.isEmpty {
-                groupedWorkouts.append(GroupedWorkouts(movement: movement, workouts: workoutsForMovement))
-            }
-        }
-        
-        return sortGroupedWorkouts(groupedWorkouts)
-    }
-    
-    private func sortGroupedWorkouts(_ groupedWorkouts: [GroupedWorkouts]) -> [GroupedWorkouts] {
-        return groupedWorkouts.sorted(by: { $0.movement.rawValue < $1.movement.rawValue })
-    }
-}
-
